Change data type of Physical Exercise to long

+++ CodeChef/Physical_Exercise.java
@@ -11,22 +11,22 @@ import java.lang.*;
 
 	    	for(int t=0;t<T;t++) {
 
-	    		int x = sc.nextInt();
-	    		int y = sc.nextInt();
+	    		long x = sc.nextLong();
+	    		long y = sc.nextLong();
 	    		int N = sc.nextInt();
 	    		int M = sc.nextInt();
 	    		int K = sc.nextInt();
 
-	    		int Narray[] = new int[N*2];
-	    		int Marray[] = new int[M*2];
-	    		int Karray[] = new int[K*2];
+	    		long Narray[] = new long[N*2];
+	    		long Marray[] = new long[M*2];
+	    		long Karray[] = new long[K*2];
 
 	    		for(int n=0;n<N*2;n++)
-	    			Narray[n] = sc.nextInt();
+	    			Narray[n] = sc.nextLong();
 	    		for(int m=0;m<M*2;m++)
-	    			Marray[m] = sc.nextInt();
+	    			Marray[m] = sc.nextLong();
 	    		for(int k=0;k<K*2;k++)
-	    			Karray[k] = sc.nextInt();
+	    			Karray[k] = sc.nextLong();
 
 	    		for(int i=0;i<N*2-1;i+=2) {
